+++ AndroidTest01/.idea/dictionaries/ryota.xml
@@ -1,0 +1,8 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="ryota">
+    <words>
+      <w>abcd</w>
+      <w>dobl</w>
+    </words>
+  </dictionary>
+</component>
