+++ AndroidTest01/.idea/inspectionProfiles/profiles_settings.xml
@@ -1,0 +1,6 @@
+<component name="InspectionProjectProfileManager">
+  <settings>
+    <option name="PROJECT_PROFILE" value="stv inspections" />
+    <option name="USE_PROJECT_PROFILE" value="false" />
+  </settings>
+</component>
