+++ AndroidTest01/app/src/test/java/com/example/ryota/androidtest01/ExampleUnitTest.java
@@ -1,17 +1,13 @@
 package com.example.ryota.androidtest01;
 
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
 /**
  * Example local unit test, which will execute on the development machine (host).
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() {
-        assertEquals(4, 2 + 2);
-    }
+class ExampleUnitTest {
+//    @Test
+//    public void addition_isCorrect() {
+//        assertEquals(4L, (long) (2 + 2));
+//    }
 }
