+++ AndroidTest01/lint.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<lint>
+    <issue id="all">
+        <ignore path="src/test/**/*" />
+        <ignore path="src/androidTest/**/*" />
+    </issue>
+</lint>
